moved event formatters under formatters (restructure)
