+++ pkg/domain/formatters/events/tenant.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package formatters
+package events
 
 import (
 	"context"
